--- client/src/app/+admin/moderation/abuse-list/abuse-list.component.scss
@@ -1,32 +1,0 @@
-@import 'mixins';
-@import 'miniature';
-
-.video-details-date-updated {
-  font-size: 90%;
-  margin-top: .1rem;
-}
-
-.video-details-links {
-  @include disable-default-a-behaviour;
-}
-
-.abuse-states .glyphicon-comment {
-  margin-left: 0.5rem;
-}
-
-.input-group {
-  @include peertube-input-group(300px);
-
-  .dropdown-toggle::after {
-    margin-left: 0;
-  }
-}
-
-.abuse-messages {
-  my-global-icon {
-    width: 22px;
-    margin-left: 3px;
-    position: relative;
-    top: -2px;
-  }
-}
