+++ client/src/app/+admin/moderation/abuse-list/index.ts
@@ -1,3 +1,1 @@
-export * from './abuse-details.component'
 export * from './abuse-list.component'
-export * from './moderation-comment-modal.component'
