+++ client/src/app/+admin/moderation/moderation.component.ts
@@ -3,7 +3,7 @@ import { ServerService } from '@app/core'
 
 @Component({
   templateUrl: './moderation.component.html',
-  styleUrls: [ './moderation.component.scss' ]
+  styleUrls: [ ]
 })
 export class ModerationComponent implements OnInit {
   autoBlockVideosEnabled = false
