+++ client/src/app/+my-account/my-account-abuses/my-account-abuses-list.component.html
@@ -3,4 +3,4 @@
   <ng-container i18n>Reports</ng-container>
 </h1>
 
-<my-abuse-list-table viewType="admin" baseRoute="/admin/moderation/abuses/list"></my-abuse-list-table>
+<my-abuse-list-table viewType="user" baseRoute="/my-account/abuses"></my-abuse-list-table>
