+++ client/src/app/+my-account/my-account-abuses/my-account-abuses-list.component.ts
@@ -1,0 +1,11 @@
+
+import { Component } from '@angular/core'
+
+@Component({
+  selector: 'my-account-abuses-list',
+  templateUrl: './my-account-abuses-list.component.html',
+  styleUrls: [ ]
+})
+export class MyAccountAbusesListComponent {
+
+}
