+++ client/src/app/+my-account/my-account-blocklist/my-account-server-blocklist.component.ts
@@ -3,7 +3,7 @@ import { BlocklistComponentType, GenericServerBlocklistComponent } from '@app/sh
 
 @Component({
   selector: 'my-account-server-blocklist',
-  styleUrls: [ '../../+admin/moderation/moderation.component.scss', '../../shared/shared-moderation/server-blocklist.component.scss' ],
+  styleUrls: [ '../../shared/shared-moderation/moderation.scss', '../../shared/shared-moderation/server-blocklist.component.scss' ],
   templateUrl: '../../shared/shared-moderation/server-blocklist.component.html'
 })
 export class MyAccountServerBlocklistComponent extends GenericServerBlocklistComponent {
