+++ client/src/app/shared/shared-abuse-list/abuse-details.component.scss
@@ -1,0 +1,34 @@
+@import 'variables';
+@import 'mixins';
+@import 'miniature';
+
+.screenratio {
+  div {
+    @include miniature-thumbnail;
+
+    display: inline-flex;
+    justify-content: center;
+    align-items: center;
+    color: pvar(--inputPlaceholderColor);
+  }
+
+  @include large-screen-ratio($selector: 'div, ::ng-deep iframe') {
+    width: 100% !important;
+    height: 100% !important;
+    left: 0;
+  };
+}
+
+.comment-html {
+  background-color: #ececec;
+  padding: 10px;
+}
+
+.abuse-details-date-updated {
+  font-size: 90%;
+  margin-top: .1rem;
+}
+
+.abuse-details-links {
+  @include disable-default-a-behaviour;
+}
