+++ client/src/app/shared/shared-abuse-list/index.ts
@@ -1,0 +1,7 @@
+export * from './abuse-message-modal.component'
+export * from './abuse-list-table.component'
+export * from './abuse-details.component'
+export * from './moderation-comment-modal.component'
+export * from './processed-abuse.model'
+
+export * from './shared-abuse-list.module'
